chore: switch max_cost to float (#44)

+++ quantinuum_schemas/models/backend_config.py
@@ -411,10 +411,10 @@ class HeliosConfig(BaseBackendConfig):
     system_name: str = "Helios-1"
     emulator_config: HeliosEmulatorConfig | None = None
 
-    max_cost: int | None = None
+    max_cost: float | None = None
 
     attempt_batching: bool = False
-    max_batch_cost: int = 2000
+    max_batch_cost: float = 2000.0
 
     options: QuantinuumOptions | None = None
